+++ crates/egui_demo_app/Cargo.toml
@@ -62,7 +62,6 @@ env_logger = "0.10"
 
 # web:
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-console_error_panic_hook = "0.1.6"
 wasm-bindgen = "=0.2.84"
 wasm-bindgen-futures = "0.4"
 web-sys = "0.3"
